pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stageguard"
version = "0.1.0"
description = "Backbone-agnostic physiological constraints for neural sleep staging"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "scikit-learn>=1.2",
    "pyyaml>=6.0",
    "h5py>=3.8",
    "pandas>=2.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.3"]

[tool.setuptools.packages.find]
include = ["stageguard*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
